Make duration conversion more robust
Unfortunately the native library returns a different format for the
duration, using the same key, if the "complete" option is set or if
it is not set.

Without "complete" it looks like "6 min 42 s", whereas with the
complete option it looks like "00:06:42.168" or possibly even like
"00:06:42.168 (00:06:42;05)".

With this change all of these formats are tried.